mo shit

+++ downloads.php
@@ -7,14 +7,6 @@
 	<body>
 	<?php
 		session_start();
-	    if($_SESSION['loggedIn'] == 1)
-		{
-			echo '<div class = "hback"><h1 class = "title">'.$_SESSION['user'].'\'s DOWNLOAD PAGE</h1></div>';
-		}
-		else
-		{
-			echo '<div class = "hback"><h1 class = "title">DOWNLOAD PAGE</h1></div>';
-		}
 		echo '<center><ul id="navbar">';
 			echo'<li><a href = "home.php">HOME</a></li>';
 			echo '<li><a href = "statistics.php">GLOBAL STATISTICS</a></li>';
@@ -23,14 +15,17 @@
 			echo '<li><a href = "login.php">LOGIN</a><li>';
 		echo '</ul></center>';
 		echo '<div class = "normal"></div>';
-		if($_SESSION['loggedIn'] == 1)
-		{
-			echo '<div><p class = "gold">'.$_SESSION['user'].'\'s DOWLOADS</p></div>';
-		}
-		else
-		{
-			echo '<div><p class = "gold">DOWLOADS</p></div>';
-		}
+        if(array_key_exists('loggedIn', $_SESSION))
+        {
+            if($_SESSION['loggedIn'] == 1)
+            {
+                echo '<div><p class = "gold">'.$_SESSION['user'].'\'s DOWLOADS</p></div>';
+            }
+            else
+            {
+                echo '<div><p class = "gold">DOWLOADS</p></div>';
+            }
+        }
 			
 			$fileText = file_get_contents('/home/geralab/pass.txt', FILE_USE_INCLUDE_PATH);
 			$dbPassword = trim($fileText);
